StyleCode after use SyleCop
