+++ Bot.cs
@@ -1,0 +1,17 @@
+﻿namespace Encapsulation_task_1
+{
+    public class Bot
+    {
+        private readonly Weapon _weapon;
+
+        public Bot(Weapon weapon)
+        {
+            _weapon = weapon;
+        }
+
+        public void OnSeePlayer(Player player)
+        {
+            _weapon.Fire(player);
+        }
+    }
+}
